refactor: streamline component resolution and remove axios bootstrap
